fix tests and fix funtional-style operators
